+++ src/onegov.feriennet/onegov/feriennet/assets/js/donation.js
@@ -1,0 +1,11 @@
+$(document).ready(function() {
+    $('.donation-layout form .button').on('mouseenter', function() {
+        $('#Eyes').attr('display', 'none');
+        $('#Pleased').removeAttr('display');
+    });
+
+    $('.donation-layout form .button').on('mouseout', function() {
+        $('#Pleased').attr('display', 'none');
+        $('#Eyes').removeAttr('display');
+    });
+});